New Rule Activities and renamed an Activity
I added a new activity for Penalties and a separate activity for player
Positions.

The PlayerPositions activity brings up a list for offensive, defensive,
and special positions. I still have to set it up to go to the
appropriate one.

For the PlayerPositions and RuleList activities, I added a switch for
each to determine the positions and bring up the appropriate
activity/view.