Added logo png files, edited team HTML files, and added pictures to HTML
I added a folder into the assets folder which contains the logos of the
teams. I found this to be the easiest way to pull the logos up in the
HTML files.

I edited all the HTML files to display the appropriate logos and
subtitles. I also added the some styles to the CSS file once again.